# public/spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MailBaby Email Delivery and Management Service API",
    "version": "1.5.0",
...
    "termsOfService": "https://www.interserver.net/terms-of-service.html",
    "contact": {
      "name": "Mail Baby",
      "url": "https://www.mail.baby/contact/",
      "email": "support@interserver.net"
    },
    "license": {
      "name": "GNU GPLv3",
      "url": "https://www.gnu.org/licenses/gpl.txt"
    }
  },
  "servers": [
    {
      "url": "https://api.mailbaby.net",
      "description": "Live API Endpoint"
    }
  ],
  "paths": {
...
    "/mail": {
      "get": {
        "tags": [
          "Services"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MailOrder"
                  }
                },
                "examples": {
                  "MailOrders": {
                    "value": [
                      {
                        "id": 39,
                        "status": "active",
                        "username": "mb39"
                      },
                      {
                        "id": 33,
                        "status": "canceled",
                        "username": "mb33",
                        "comment": "test order"
                      }
                    ]
                  }
                }
              }
            },
            "links": {
              "getMailOrderById": {
                "operationId": "getMailOrderById",
                "parameters": {
                  "id": "$response.body#/0/id"
                },
                "description": "The `id` value returned in the response can be used as the `id` path parameter in `GET /mail/{id}` to retrieve full order details."
              },
              "sendMailByIdLink": {
                "operationId": "sendMail",
                "parameters": {
                  "id": "$response.body#/0/id"
                },
                "description": "The `id` value returned in the response can be used as the `id` body parameter in `POST /mail/send`."
              },
              "sendAdvMailByIdLink": {
                "operationId": "sendAdvMail",
                "parameters": {
                  "id": "$response.body#/0/id"
                },
                "description": "The `id` value returned in the response can be used as the `id` body parameter in `POST /mail/advsend`."
              },
              "viewMailLogByIdLink": {
                "operationId": "viewMailLog",
                "parameters": {
                  "id": "$response.body#/0/id"
                },
                "description": "The `id` value returned in the response can be used as the `id` query parameter in `GET /mail/log`."
              },
              "getStatsByIdLink": {
                "operationId": "getStats",
                "parameters": {
                  "id": "$response.body#/0/id"
                },
                "description": "The `id` value returned in the response can be used as the `id` query parameter in `GET /mail/stats`."
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "operationId": "getMailOrders",
        "summary": "List your provisioned mail orders (sending accounts)",
        "description": "Use this **before sending** if you need to discover which mail order `id` to pass to `/mail/send`, `/mail/advsend`, `/mail/rawsend`, `/mail/log`, or `/mail/stats`. Returns every mail order (active **and** canceled) on your account. Each record includes the numeric `id`, the `status` (`active` or `canceled`), the SMTP `username` (always `mb<id>`), and an optional `comment`.\n\nThe `id` values returned here are used as the `id` input parameter on all sending endpoints (`/mail/send`, `/mail/advsend`, `/mail/rawsend`) as well as the log and stats queries.  When the `id` parameter is omitted on those calls the API automatically picks the **first active** order returned by this endpoint.\n\nTo retrieve full details \u2014 including the current SMTP password \u2014 for a single order use `GET /mail/{id}`.\n"
      }
    },
    "/mail/{id}": {
      "get": {
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "example": 39,
            "name": "id",
            "description": "The numeric ID of the mail order.",
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailOrderDetail"
                },
                "examples": {
                  "MailOrderDetail": {
                    "value": {
                      "id": 39,
                      "status": "active",
                      "username": "mb39",
                      "password": "s3cr3tpassword"
                    }
                  }
                }
              }
            },
            "links": {
              "sendMailByIdLink": {
                "operationId": "sendMail",
                "parameters": {
                  "id": "$response.body#/id"
                },
                "description": "The `id` value can be used as the `id` body parameter in `POST /mail/send`."
              },
              "sendAdvMailByIdLink": {
                "operationId": "sendAdvMail",
                "parameters": {
                  "id": "$response.body#/id"
                },
                "description": "The `id` value can be used as the `id` body parameter in `POST /mail/advsend`."
              },
              "viewMailLogByIdLink": {
                "operationId": "viewMailLog",
                "parameters": {
                  "id": "$response.body#/id"
                },
                "description": "The `id` value can be used as the `id` query parameter in `GET /mail/log`."
              },
              "getStatsByIdLink": {
                "operationId": "getStats",
                "description": "Use `GET /mail/stats` to view sending statistics for this mail order."
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/BadInput"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "operationId": "getMailOrderById",
        "summary": "Get one mail order's full detail including SMTP password",
        "description": "**Returns the SMTP password** in addition to the basic `GET /mail` fields. Use when you need to authenticate against `relay.mailbaby.net:25` from an SMTP client directly. Required: `id` (path; obtain from `GET /mail`). Returns the full detail record for one specific mail order identified by its numeric `id`.\n\nThe `username` and `password` values returned here can be used directly to authenticate against `relay.mailbaby.net:25` (SMTP AUTH) if you need to send email via a native SMTP client rather than through the REST API.\n\nThe `id` path parameter is the same integer `id` value returned by `GET /mail`.\n"
      }
    },
    "/mail/send": {
      "post": {
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/SendMail"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMail"
              }
            }
          },
          "required": true
        },
        "tags": [
          "Sending"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "examples": {
                  "SendSuccess": {
                    "value": {
                      "status": "ok",
                      "text": "185caa69ff7000f47c"
                    }
                  }
                }
              }
            },
            "links": {
              "lookupByTransactionId": {
                "operationId": "viewMailLog",
                "parameters": {
                  "mailid": "$response.body#/text"
                },
                "description": "The transaction ID in `text` can be passed as `mailid` to `GET /mail/log` to retrieve the delivery record."
              }
            },
            "description": "Email accepted for delivery"
          },
          "400": {
            "$ref": "#/components/responses/BadInput"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "operationId": "sendMail",
        "summary": "Send a simple email (single recipient, plain or HTML)",
        "description": "**Use this for the simplest case** \u2014 one or a few plain-string recipients, no attachments, no CC/BCC, no named contacts. Required: `from`, `to`, `subject`, `body`. Optional: `id` (selects which mail order to send through; auto-selects your first active order if omitted). Sends through one of your mail orders using a flat set of fields.\n\n**HTML detection** is automatic \u2014 if the `body` value contains any HTML tags the message will be sent as `text/html`; otherwise it is sent as `text/plain`.\n\nThe `from` address is also automatically set as the `Reply-To` header.\n\n*Note*: If you need to send to multiple recipients, add CC/BCC, or include file attachments, use `POST /mail/advsend` instead.  If you have a pre-built RFC 822 message (e.g. already DKIM-signed), use `POST /mail/rawsend`.\n\nOn success the response `text` field contains the relay transaction ID.  This ID can be used with `GET /mail/log` (via the `mailid` query parameter) to look up the delivery record.\n"
      }
    },
    "/mail/rawsend": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMailRaw"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SendMailRaw"
              }
            }
          },
          "required": true
        },
        "tags": [
          "Sending"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "examples": {
                  "RawSendSuccess": {
                    "value": {
                      "status": "ok",
                      "text": "185caa69ff7000f47c"
                    }
                  }
                }
              }
            },
            "links": {
              "lookupByTransactionId": {
                "operationId": "viewMailLog",
                "parameters": {
                  "mailid": "$response.body#/text"
                },
                "description": "The transaction ID in `text` can be passed as `mailid` to `GET /mail/log` to retrieve the delivery record."
              }
            },
            "description": "Email accepted for delivery"
          },
          "400": {
            "$ref": "#/components/responses/BadInput"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "operationId": "rawMail",
        "summary": "Send a pre-built RFC 822 message verbatim (preserves DKIM)",
        "description": "**Use this when you have already built (and possibly DKIM-signed) the full RFC 822 message yourself** and want the relay to transmit it byte-for-byte. Required: `raw_email` (full RFC 822 string with headers + body). Optional: `id` (mail order to authenticate as). Accepts a complete, pre-built RFC 822 email message (headers + body) as a string and injects it into the relay without any modification.\n\nThis endpoint is particularly useful when the message has already been **DKIM-signed** \u2014 because the relay transmits the exact bytes you provide, the DKIM signature remains intact. If you use the other sending endpoints the relay may add or reorder headers, breaking an existing signature.\n\nThe `From` and recipient addresses are parsed automatically from the message headers (`From`, `To`, `Cc`, `Bcc`).  You do **not** need to specify them separately.\n\nIf an `id` is provided it must correspond to an active mail order on your account. If omitted, the first active order is selected automatically.  The SMTP credentials for the selected order are used to authenticate with the relay.\n\nOn success the response `text` field contains the relay transaction ID.  This ID can be used with `GET /mail/log` (via the `mailid` query parameter) to look up the delivery record.\n"
      }
    },
    "/mail/advsend": {
      "post": {
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/SendMailAdv"
              },
              "examples": {
                "BasicForm": {
                  "summary": "Simple string addresses",
                  "value": {
                    "subject": "Welcome",
                    "body": "Hello",
                    "from": "user@domain.com",
                    "to": "support@interserver.net"
                  }
                },
                "NamedForm": {
                  "summary": "RFC 822 named addresses",
                  "value": {
                    "subject": "Welcome",
                    "body": "Hello",
                    "from": "Joe <user@domain.com>",
                    "to": "Jane <support@interserver.net>"
                  }
                },
                "AdvSendExample": {
                  "summary": "Structured array with attachment",
                  "value": {
                    "subject": "Welcome",
                    "body": "Hello",
                    "from": {
                      "email": "user@domain.com"
                    },
                    "to": [
                      {
                        "email": "someone@client.com",
                        "name": "Mr Client"
                      }
                    ],
                    "attachments": [
                      {
                        "filename": "message.txt",
                        "data": "aGVsbG8gdGhlcmUK"
                      }
                    ],
                    "id": 66
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMailAdv"
              },
              "examples": {
                "BasicJson": {
                  "summary": "Plain string addresses",
                  "value": {
                    "subject": "Welcome",
                    "body": "Hello",
                    "from": "user@domain.com",
                    "to": "support@interserver.net"
                  }
                },
                "ArrayJson": {
                  "summary": "Structured objects",
                  "value": {
                    "subject": "Welcome",
                    "body": "Hello",
                    "from": {
                      "name": "Joe",
                      "email": "user@domain.com"
                    },
                    "to": [
                      {
                        "name": "Joe",
                        "email": "support@interserver.net"
                      }
                    ]
                  }
                },
                "NameEmailJson": {
                  "summary": "RFC 822 named string",
                  "value": {
                    "subject": "Welcome",
                    "body": "Hello",
                    "from": "Joe <user@domain.com>",
                    "to": "Joe <support@interserver.net>"
                  }
                },
                "FullExample": {
                  "summary": "Full-featured with CC, attachment, id",
                  "value": {
                    "subject": "Welcome",
                    "body": "Hello",
                    "from": {
                      "email": "user@domain.com",
                      "name": "Sender Name"
                    },
                    "to": [
                      {
                        "email": "someone@client.com",
                        "name": "Mr Client"
                      }
                    ],
                    "cc": [
                      {
                        "email": "someone2@client.com",
                        "name": "Mr Second"
                      }
                    ],
                    "bcc": [
                      {
                        "email": "archive@company.com"
                      }
                    ],
                    "replyto": [
                      {
                        "email": "replies@domain.com"
                      }
                    ],
                    "attachments": [
                      {
                        "filename": "message.txt",
                        "data": "aGVsbG8gdGhlcmUK"
                      }
                    ],
                    "id": 66
                  }
                }
              }
            }
          },
          "required": true
        },
        "tags": [
          "Sending"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "examples": {
                  "SendSuccess": {
                    "value": {
                      "status": "ok",
                      "text": "185caa69ff7000f47c"
                    }
                  }
                }
              }
            },
            "links": {
              "lookupByTransactionId": {
                "operationId": "viewMailLog",
                "parameters": {
                  "mailid": "$response.body#/text"
                },
                "description": "The transaction ID in `text` can be passed as `mailid` to `GET /mail/log` to retrieve the delivery record."
              }
            },
            "description": "Email accepted for delivery"
          },
          "400": {
            "$ref": "#/components/responses/BadInput"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "operationId": "sendAdvMail",
        "summary": "Send email with multiple recipients, CC/BCC, attachments, named contacts",
        "description": "**Use this when the simple send is not enough** \u2014 multiple recipients, CC/BCC, named contacts, custom Reply-To, or file attachments. Required: `from`, `to`, `subject`, `body`. Optional: `cc`, `bcc`, `replyto`, `attachments` (array of `{filename,data}` with base64 data), `id`. Sends through one of your mail orders with full control over recipients, headers, and attachments.\n\n**Content-type flexibility** \u2014 the request body may be submitted as either `application/x-www-form-urlencoded` or `application/json`.  When using form encoding, address fields (`from`, `to`, `replyto`, `cc`, `bcc`) accept both a plain RFC 822 comma-separated string (e.g. `\"Joe <joe@example.com>, jane@example.com\"`) and the structured `[{\"email\":\"...\",\"name\":\"...\"}]` array format.  When using JSON, both formats are equally supported.\n\n**HTML detection** is automatic \u2014 if `body` contains HTML tags the `Content-Type` is set to `text/html`; otherwise `text/plain`.\n\n**Attachments** must be base64-encoded.  Each attachment requires at least a `data` field; the optional `filename` field controls the attachment name shown to recipients.\n\nIf an `id` is provided it must correspond to an active mail order on your account. If omitted, the first active order is selected automatically.\n\nOn success the response `text` field contains the relay transaction ID which can be looked up in `GET /mail/log` via the `mailid` query parameter.\n\n## Examples\n\n### Form \u2014 basic string addresses\n```sh curl -X POST https://api.mailbaby.net/mail/advsend \\\n  -H 'X-API-KEY: YOUR_API_KEY' \\\n  -H 'Content-Type: application/x-www-form-urlencoded' \\\n  -d 'subject=Welcome' \\\n  -d 'body=Hello' \\\n  -d from=user@domain.com \\\n  -d to=support@interserver.net\n```\n### Form \u2014 RFC 822 named addresses\n```sh curl -X POST https://api.mailbaby.net/mail/advsend \\\n  -H 'X-API-KEY: YOUR_API_KEY' \\\n  -H 'Content-Type: application/x-www-form-urlencoded' \\\n  -d 'subject=Welcome' \\\n  -d 'body=Hello' \\\n  -d 'from=Joe <user@domain.com>' \\\n  -d 'to=Jane <support@interserver.net>'\n```\n### Form \u2014 multiple recipients as array\n```sh curl -X POST https://api.mailbaby.net/mail/advsend \\\n  -H 'X-API-KEY: YOUR_API_KEY' \\\n  -H 'Content-Type: application/x-www-form-urlencoded' \\\n  -d 'subject=Welcome' -d 'body=Hello' \\\n  -d from=user@domain.com \\\n  -d 'to[0][name]=Joe' -d 'to[0][email]=support@interserver.net' \\\n  -d 'to[1][name]=Jane' -d 'to[1][email]=jane@interserver.net'\n```\n### JSON \u2014 structured objects\n```sh curl -X POST https://api.mailbaby.net/mail/advsend \\\n  -H 'X-API-KEY: YOUR_API_KEY' \\\n  -H 'Content-Type: application/json' \\\n  -d '{\n    \"subject\": \"Welcome\",\n    \"body\": \"<h1>Hello</h1>\",\n    \"from\": {\"email\": \"user@domain.com\", \"name\": \"Sender Name\"},\n    \"to\": [{\"email\": \"joe@client.com\", \"name\": \"Joe Client\"}],\n    \"cc\": [{\"email\": \"manager@client.com\"}],\n    \"attachments\": [{\"filename\": \"report.pdf\", \"data\": \"BASE64_DATA\"}]\n  }'\n```\n"
      }
    },
...
    "/mail/stats": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "time",
            "description": "The time window to scope `received`, `sent`, and `volume` statistics. Does not affect `usage` or `cost`, which are always calculated over the current billing cycle. Defaults to `1h`.",
            "schema": {
              "default": "1h",
              "enum": [
                "all",
                "billing",
                "month",
                "7d",
                "24h",
                "day",
                "1h"
              ],
              "x-enum-varnames": [
                "All",
                "Billing",
                "Month",
                "SevenDays",
                "TwentyFourHours",
                "Day",
                "OneHour"
              ],
              "type": "string"
            },
            "in": "query",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailStatsType"
                },
                "examples": {
                  "StatsExample": {
                    "value": {
                      "time": "all",
                      "usage": 55,
                      "currency": "USD",
                      "cost": 1.02,
                      "received": 508,
                      "sent": 495,
                      "volume": {
                        "to": {
                          "client@domain.com": 395,
                          "user@site.net": 57
                        },
                        "from": {
                          "billing@somedomain.com": 369
                        },
                        "ip": {
                          "1.1.1.1": 142,
                          "2.2.2.2": 132
                        }
                      }
                    }
                  }
                }
              }
            },
            "links": {
              "viewMailLogLink": {
                "operationId": "viewMailLog",
                "description": "Use `GET /mail/log` to view individual message details behind the aggregate statistics."
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "operationId": "getStats",
        "summary": "Account usage, cost, and volume statistics for a time window",
        "description": "**Use this for dashboards or billing-cycle reviews.** Returns aggregate sending statistics for your mail account(s). The `time` query param controls the window for `received`, `sent`, and `volume` (`1h`, `24h`, `7d`, `month`, `day`, `billing`, `all`). The `usage` and `cost` fields are always for the current billing cycle, regardless of `time`.\n\nThe response includes:\n- **`usage`** \u2014 total messages accepted by the relay during the current billing\n  cycle (used for cost calculation).\n- **`cost`** \u2014 estimated cost for the billing cycle based on the base plan price\n  plus per-email charges.\n- **`received`** / **`sent`** \u2014 count of messages accepted by the relay /\n  successfully delivered to the destination MX within the selected `time` window.\n- **`volume`** \u2014 top-500 breakdown of message counts grouped by source IP (`ip`),\n  destination address (`to`), and sender address (`from`) within the selected window.\n\n\n**Time windows** (controlled by the `time` parameter):\n| Value | Window | |-------|--------| | `1h` | Last 1 hour (default) | | `24h` | Last 24 hours | | `7d` | Last 7 days | | `month` | Current calendar month (1st to now) | | `day` | Today (midnight to now) | | `billing` | Current billing cycle (last invoice date to next invoice date) | | `all` | All time |\n"
      }
    },
    "/mail/blocks": {
      "get": {
        "tags": [
          "Blocking"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlocks"
                },
                "examples": {
                  "MailBlocksExample": {
                    "value": {
                      "local": [
                        {
                          "date": "2023-08-07",
                          "from": "user@domain.com",
                          "messageId": "pFaRqFUEWkucjhTuIzYuoAgWU@domain.com",
                          "subject": "Test Email",
                          "to": "['client@site.com']"
                        }
                      ],
                      "mbtrap": [
                        {
                          "date": "2023-08-07",
                          "from": "user@domain.com",
                          "messageId": "pFaRqFUEWkucjhTuIzYuoAgWU@domain.com",
                          "subject": "Test Email",
                          "to": "['client@site.com']"
                        }
                      ],
                      "subject": [
                        {
                          "from": "user@domain.com",
                          "subject": "Test Email"
                        }
                      ]
                    }
                  }
                }
              }
            },
            "links": {
              "delistBlockLink": {
                "operationId": "delistBlock",
                "parameters": {
                  "email": "$response.body#/local/0/from"
                },
                "description": "The `from` address in any block list entry can be passed as the `email` field to `POST /mail/blocks/delete` to delist it."
              }
            },
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "operationId": "getMailBlocks",
        "summary": "List addresses currently flagged by spam-filter block lists",
        "description": "**Use this to discover which addresses to delist.** The `from` value of any returned entry can be passed to `POST /mail/blocks/delete` (as the `email` field) to remove the block. Returns addresses and messages that have been flagged by the spam filtering system for your mail account(s). Three categories are returned:\n\n- **`local`** \u2014 messages flagged by the `LOCAL_BL_RCPT` rspamd rule.  These are\n  messages sent to recipients on your account's local block list.\n- **`mbtrap`** \u2014 messages flagged by the `MBTRAP` rspamd rule.  These are messages\n  that triggered MailBaby's internal trap / honeypot detection.\n- **`subject`** \u2014 senders whose recent messages contain spam-indicative subjects\n  (strings containing `@`, `smtp`, `socks4`, or `socks5`) with high repetition\n  (more than 4 identical subjects from the same sender in the last 3 days).\n\n\nThe `local` and `mbtrap` results cover the last 5 days.  The `subject` results cover the last 3 days.\n\nA sender address returned in any of these lists can be delisted using `POST /mail/blocks/delete` with the `email` field set to that address.\n"
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadInput": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "examples": {
              "BadInputError": {
                "value": {
                  "code": 400,
                  "message": "There was a problem with the input parameters."
                }
              }
            }
          }
        },
        "description": "Bad request \u2014 one or more input parameters were missing or invalid."
      },
...
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "examples": {
              "UnauthorizedError": {
                "value": {
                  "code": 401,
                  "message": "API key is missing or invalid"
                }
              }
            }
          }
        },
        "description": "Authentication failed.  Ensure you are sending a valid `X-API-KEY` header. Obtain your API key from [my.interserver.net/account_security](https://my.interserver.net/account_security)."
      }
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "description": "Pass your API key in the `X-API-KEY` HTTP request header.  Obtain your key from the [Account Security](https://my.interserver.net/account_security) page on my.interserver.net.",
        "name": "X-API-KEY",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ],
...
}